--- src/data_access/__init__.py
@@ -1,1 +1,0 @@
-"""Data access helpers for dashboard table loading and refresh."""
