+++ src/social_media_insights.py
@@ -1,5 +1,3 @@
-"""Aggregations for the Social Media Analysis page (no DB changes; derived in-app)."""
-
 from __future__ import annotations
 
 from typing import Any
