+++ src/social_media_plots.py
@@ -1,5 +1,3 @@
-"""Plotly figures for social media pages (no database import)."""
-
 from __future__ import annotations
 
 import numpy as np
