--- src/table_columns/__init__.py
@@ -1,1 +1,0 @@
-"""Dash DataTable column definitions."""
