+++ tests/unit/test_css_tokens_smoke.py
@@ -1,5 +1,3 @@
-"""Phase 7e - assert core design tokens and shell hooks exist in global CSS."""
-
 from pathlib import Path
