+++ tests/unit/test_schedule_page_helpers.py
@@ -1,5 +1,3 @@
-"""Unit tests for schedule page helpers and tonight's card builder."""
-
 from __future__ import annotations
 
 from unittest.mock import patch
